refactor(math): move math from utils
