+++ mdbenchmark/utils.py
@@ -88,19 +88,6 @@ def retrieve_host_template(host=None):
     return ENV.get_template(host)
 
 
-def lin_func(x, m, b):
-    return m * x + b
-
-
-def calc_slope_intercept(x, y):
-    x = np.asarray(x)
-    y = np.asarray(y)
-    diff = x - y
-    slope = diff[1] / diff[0]
-    intercept = x[1] - (x[0] * slope)
-    return np.hstack([slope, intercept])
-
-
 def guess_ncores():
     """Guess the number of physical CPU cores.
